chore: prealpha 7 world time is now passed as a `long` added `phase_shift` to orbital position config

+++ src/main/java/dev/galacticraft/api/universe/position/CelestialPosition.java
@@ -28,11 +28,11 @@ import dev.galacticraft.api.registry.AddonRegistry;
 public record CelestialPosition<C extends CelestialPositionConfig, T extends CelestialPositionType<C>>(T type, C config) {
     public static final Codec<CelestialPosition<?, ?>> CODEC = AddonRegistry.CELESTIAL_POSITION_TYPE.dispatch(CelestialPosition::type, CelestialPositionType::codec);
 
-    public double x(int worldTime, float delta) {
+    public double x(long worldTime, float delta) {
         return this.type().x(this.config(), worldTime, delta);
     }
 
-    public double y(int worldTime, float delta) {
+    public double y(long worldTime, float delta) {
         return this.type().y(this.config(), worldTime, delta);
     }
